expand mtd update bash script
some targets need specific commands to be called before
and after update is done. gpio's might need to be asserted
or drivers loaded. there is no one set of the specific
commands so introduce calls to setup_$TARGET and
cleanup_$TARGET bash scripts before and after update
is performed.
Also add messaging to fw update Tasks so the user can see
if update is happening and what is happening with it.

Signed-off-by: Radivoje (Ogi) Jovanovic <rjovanovic@nvidia.com>
fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-12149